+++ src/templating.ts
@@ -1,6 +1,7 @@
 const FILE_COUNT_EXP = /\{\{\s*fileCount\s*\}\}/;
 const FILES_LIST_EXP = /\{\{\s*filesList\s*\}\}/;
 const OPTIMIZED_COUNT_EXP = /\{\{\s*optimizedCount\s*\}\}/;
+const SKIPPED_COUNT_EXP = /\{\{\s*skippedCount\s*\}\}/;
 const SVG_COUNT_EXP = /\{\{\s*svgCount\s*\}\}/;
 
 const format = {
@@ -12,6 +13,9 @@ const format = {
   },
   optimizedCount: (template: string, value: number): string => {
     return template.replace(OPTIMIZED_COUNT_EXP, value.toString());
+  },
+  skippedCount: (template: string, value: number): string => {
+    return template.replace(SKIPPED_COUNT_EXP, value.toString());
   },
   svgCount: (template: string, value: number): string => {
     return template.replace(SVG_COUNT_EXP, value.toString());
@@ -34,10 +38,11 @@ function formatAll(
 
 
 export type CommitData = {
-  fileCount: number;
-  filePaths: string[];
-  optimizedCount: number;
-  svgCount: number;
+  readonly fileCount: number;
+  readonly filePaths: string[];
+  readonly optimizedCount: number;
+  readonly skippedCount: number;
+  readonly svgCount: number;
 }